requirements.txt
Markdown==3.6
PyYAML==6.0.1
reportlab==4.2.0